+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.appspot.pistatium.dotstream">
+    package="com.appspot.pistatium.dotsubscribe">
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.appspot.pistatium.dotsubscribe.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
