+++ app/src/main/kotlin/com/appspot/pistatium/dotsubscribe/models/Article.kt
@@ -1,4 +1,4 @@
-package com.appspot.pistatium.dotstream.models
+package com.appspot.pistatium.dotsubscribe.models
 
 
 data class Article(val title: String, val url: String)
