+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="contentViewModel"
-            type="com.appspot.pistatium.dotstream.viewmodels.ContentViewModel" />
+            type="com.appspot.pistatium.dotsubscribe.viewmodels.ContentViewModel" />
     </data>
 
     <RelativeLayout
