+++ app/src/main/res/layout/article_cell.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="article"
-            type="com.appspot.pistatium.dotstream.models.Article" />
+            type="com.appspot.pistatium.dotsubscribe.models.Article" />
     </data>
     <RelativeLayout
         android:layout_margin="5dp"
