+++ app/src/main/res/layout/content_view.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.appspot.pistatium.dotstream.viewmodels.ContentViewModel" />
+            type="com.appspot.pistatium.dotsubscribe.viewmodels.ContentViewModel" />
     </data>
     <RelativeLayout
         android:layout_width="match_parent"
