+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">DotStream</string>
+    <string name="app_name">DotSubscribe</string>
 </resources>
